feat: add some tooling for security analysis
