+++ air/Cargo.toml
@@ -25,7 +25,6 @@ fri = { version = "0.13", path = "../fri", package = "winter-fri", default-featu
 libm = "0.2"
 math = { version = "0.13", path = "../math", package = "winter-math", default-features = false }
 utils = { version = "0.13", path = "../utils/core", package = "winter-utils", default-features = false }
-libc-print = "0.1.23"
 
 [dev-dependencies]
 rand-utils = { version = "0.13", path = "../utils/rand", package = "winter-rand-utils" }
