+++ air/src/lib.rs
@@ -32,6 +32,9 @@
 #[macro_use]
 extern crate alloc;
 
+#[cfg(feature = "std")]
+extern crate std;
+
 pub mod proof;
 
 mod errors;
